Skip 'Ping npm registry' task when --no-publish
Skip the task 'Ping npm registry' when the option `--no-publish` is
present on the command line, as this task is only useful when publishing
a package.

I tried to run `np --no-publish` with a custom npm registry ([Nexus](https://help.sonatype.com/display/NXRM3/Node+Packaged+Modules+and+npm+Registries)) set in `~/.npmrc` and this fails without this patch as it tries to ping the npm registry, which is unfortunately [not supported on Nexus](https://issues.sonatype.org/browse/NEXUS-13434).