linkify commits: repository can be undefined
Trying the new version of np on a project that doesn't define
the repository entry in `package.json` results in this error

> Cannot read property 'url' of undefined

This is related to commit 263b0c3c92365525c2b1887b069a6ca9d0c508ef
(linkify commits) which assumes that the repository entry is always present
in `package.json`, which is not always the case.

This simply check that `pkg.repository` is defined before using
`pkg.repository.url`.