🐛  Fix #600: Unexpected end of json input
NPM is mixing json with human readable error messages (npm/cli#2740)
and refuse to fix it because they want more time to plan (npm/cli#3437).
This provides a temporary fix by simply filtering out every line that starts
with `npm ERR!` so we have a clean JSON input.

As a side effect, this will be broken when attaching a debugger to the
process, but that would have been broken without this fix.